Initial code structure
The Breakout class will have the main game loop and most of the logic -
much of it will probably need to be split out into separate classes as
we develop. The objects and Commands are pretty well defined. Only the
Loader knows about the different types of Commands; the main game logic
just treats them all as Commandables.

In addition to doing the main game loop, we need Brick and its
associated Command and the Buttons. Keyboard input also needs to be
added in a nice way. Any object that wants keyboard input should
probably be an Observer and register with some Observable that actually
gets the input.