Add collision detection for ball
This required giving ball some way of accessing the size of the board. I
opted to do this by changing the Moveable interface to pass in a
Dimension, which required changing Commandable, BallMoveCommand,
PaddleMoveCommand, and Paddle.move(). Passing things through the execute
method of a Command seems wrong, but I'm not sure how else to do it
other than making things globally accessible which feels even more
wrong.