Add undo/reset, KeyListener functionality, and some collision fixes.
Should have split this up but I got carried away. undo/reset is mostly
contained within Breakout and just manipulates the commandGroupStack.
After the paddle
could
move,
if
it
was
moving
into the ball when they collided, the ball would get stuck inside the
paddle. I fixed this by basically saying that when the ball collides
with an object it can't collide with that object again for some number
of frames.